Add a strict mode to the JSON parser.
This will mainly be used by the Configurable trait, but it disables the functionality that tries to assume if a string is an object or array if the boundaries are missing around the root element, and simply treats it as a string. This means that some strings (ie. "MySQL failed, please check query") won't be interpreted as an array.