Change mounting and template definition
This creates more flexibility on mounting, by allowing a developer to explicity provide or not provide a mount point and (optionally) a template.

If a mount point is not provided, mounting is deferred until the developer chooses to. Otherwise, the mount point either acts as a container or as the reactive element, depending on the template.