Add Snowboard tear down capability
Tearing down Snowboard will destruct all plugins and remove all Snowboard functionality. It's mainly for testing, but can be useful for people who want to remove Snowboard and re-initialise in circumstances.

PluginBase now also contains a "destructed" boolean to track if it's destructed. Since objects cannot delete themselves entirely in JS, this is the next best thing. Garbage collection should remove the objects before there's any need to query this boolean, but again, this is useful for tests.