Remove instanceof checks
Due to the way that build processes work, sometimes using "instanceof" checks with abstracts causes them to fail because of modifications to the prototype chain.

Given that, at the end of the day, JavaScript is loose enough that if it quacks like a duck, it is a duck (ie. as long as the API is the same, it's good enough), we'll just simply check for the existence of all PluginBase methods in a class in order to determine if something is a plugin.

Users will still need to extend PluginBase or Singleton for proper framework functionality to occur, but at least if there's a prototype chain mismatch, it won't be a fail state.